requirements.txt
langchain==0.2.11
langchain-community==0.2.10
langchain-core==0.2.28
langchain-text-splitters==0.2.2
sentence-transformers
transformers
accelerate
pinecone-client==3.2.2
pypdf
docx2txt
tiktoken
fastapi
uvicorn
python-dotenv